Move onComment and onIssue to endpoints folder
