+++ endpoints/onIssue.ts
@@ -10,9 +10,9 @@ import { IApiRequest, IApiResponse } from '@rocket.chat/apps-engine/definition/a
 import { ApiEndpoint } from '@rocket.chat/apps-engine/definition/api/ApiEndpoint';
 import { IApiEndpointInfo } from '@rocket.chat/apps-engine/definition/api/IApiEndpointInfo';
 
-import { IssueEventEnum } from './enums/IssueEventEnum';
-import { parseJiraDomainFromIssueUrl, startNewMessageWithDefaultSenderConfig } from './lib/helpers';
-import { getConnectedProjects } from './lib/persistence';
+import { IssueEventEnum } from '../enums/IssueEventEnum';
+import { parseJiraDomainFromIssueUrl, startNewMessageWithDefaultSenderConfig } from '../lib/helpers';
+import { getConnectedProjects } from '../lib/persistence';
 
 export class OnIssueEndpoint extends ApiEndpoint {
     public path: string = 'on_issue';
